blog/requirement.txt
asgiref==3.2.10
certifi==2020.6.20
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-mdeditor==0.1.18
django-reversion==3.0.7
django-stdimage==5.1.1
djangorestframework==3.11.1
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
progressbar2==3.52.1
PyMySQL==0.10.0
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
six==1.10.0
sqlparse==0.3.1
tablib==2.0.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
